Move overlapping segment iterator into fastSmoothV2.h
